+++ bower-dist/lforms.js
@@ -2615,17 +2615,8 @@ LForms.Util = {
    *  For FHIR applications, provides FHIR context information that might be
    *  needed in rendering a Quesitonnaire.
    * @param fhirContext an optional object for accessing a FHIR context and
-   *  FHIR API.  It should define the following operations:
-   *  - getCurrent(typeList, callback):  "typeList" should be a list of desired
-   *    FHIR resource types for which there is conceptually a "current" on in
-   *    the FHIR context (e.g., Patient, or Practitioner).  Only one resource
-   *    from the requested list will be returned, and the result will be null if
-   *    none of the requested resource types are available.  Because retrieving
-   *    the resource will generally be an asynchronous operation, the resource
-   *    will be returned via the first argument to the provided "callback"
-   *    function.
-   *  - getFHIRAPI():  Should return an instance of fhir.js for interacting with
-   *    the FHIR server.
+   *  a FHIR API.  It should be an instance of 'client-js', a.k.a. npm package fhirclient,
+   *  version 2.  (See http://docs.smarthealthit.org/client-js).
    */
   setFHIRContext: function setFHIRContext(fhirContext) {
     LForms.fhirContext = fhirContext;
@@ -2674,9 +2665,9 @@ LForms.Util = {
     if (!LForms._serverFHIRReleaseID) {
       // Retrieve the fhir version
       try {
-        var fhirAPI = LForms.fhirContext.getFHIRAPI();
-        fhirAPI.conformance({}).then(function (res) {
-          var fhirVersion = res.data.fhirVersion;
+        var fhirAPI = LForms.fhirContext; //fhirAPI.request('metadata?_elements=fhirVersion').then(function(res) // causes an error on lforms-smart-fhir (TBD)
+
+        fhirAPI.getFhirVersion().then(function (fhirVersion) {
           LForms._serverFHIRReleaseID = LForms.Util._fhirVersionToRelease(fhirVersion);
           console.log('Server FHIR version is ' + LForms._serverFHIRReleaseID + ' (' + fhirVersion + ')');
           callback(LForms._serverFHIRReleaseID);
@@ -3339,9 +3330,9 @@ LForms.Util = {
   },
 
   /**
-   * Get a list of warning messages about answer lists, which should have been 
+   * Get a list of warning messages about answer lists, which should have been
    * loaded from the URL in answerValueSet but were not.
-   * 
+   *
    * @param {*} formDataSource Optional.  Either the containing HTML element that
    *  includes the LForm's rendered form, a CSS selector for that element, an
    *  LFormsData object, or an LForms form definition (parsed).  If not
@@ -4990,13 +4981,18 @@ function _typeof(obj) { "@babel/helpers - typeof"; if (typeof Symbol === "functi
       var pendingPromises = []; // launchContext
 
       var contextItems = LForms.Util.findObjectInArray(this.extension, 'url', "http://hl7.org/fhir/StructureDefinition/questionnaire-launchContext", 0, true);
+      var supportedContexts = {
+        patient: 1,
+        user: 1,
+        encounter: 1
+      };
 
       var _loop = function _loop() {
         var contextItemExt = contextItems[i].extension;
-        var name = void 0;
-        var typeList = [];
+        var name = null,
+            type = null;
 
-        for (j = 0, jLen = contextItemExt.length; j < jLen; ++j) {
+        for (j = 0, jLen = contextItemExt.length; j < jLen && !(name && type); ++j) {
           fieldExt = contextItemExt[j];
 
           if (!name && fieldExt.url === 'name') {
@@ -5004,24 +5000,32 @@ function _typeof(obj) { "@babel/helpers - typeof"; if (typeof Symbol === "functi
 
             _this._checkFHIRVarName(name); // might throw
 
-          } else if (fieldExt.url === 'type') typeList.push(fieldExt.valueCode);
+          } else if (fieldExt.url === 'type') {
+            var typeCode = fieldExt.valueCode;
+            if (supportedContexts[typeCode]) type = typeCode;
+          }
         }
 
-        if (name && typeList.length > 0) {
+        if (name && type) {
           pendingPromises.push(new Promise(function (resolve, reject) {
-            // Enforce that this is truly asynchronous with setTimeout.
-            // Some implementations of getCurrent (e.g in testing) might be
-            // synchronous.
-            setTimeout(function () {
-              try {
-                LForms.fhirContext.getCurrent(typeList, function (resource) {
-                  if (resource) lfData._fhirVariables[name] = resource;
-                  resolve();
-                });
-              } catch (e) {
-                reject(e);
-              }
-            }, 1);
+            var contextResource = LForms.fhirContext[type];
+
+            if (!contextResource.id) {
+              console.error('A context resource of type ' + type + ' was requested, ' + 'but none was available'); // The loading of this resource should not be critical for the
+              // Questionnaire, because it is just for prepopulation.  Don't
+              // reject the promise.
+
+              resolve();
+            } else {
+              contextResource.read().then(function (resource) {
+                if (resource) lfData._fhirVariables[name] = resource;
+                resolve();
+              }, function fail(reason) {
+                console.log('A context resource of type ' + type + ' was requested, ' + 'but could not be read.');
+                console.error(reason);
+                resolve(); // per above, we are not rejecting the promise
+              });
+            }
           }));
         }
       };
@@ -5187,15 +5191,11 @@ function _typeof(obj) { "@babel/helpers - typeof"; if (typeof Symbol === "functi
 
               itemCodeSystem = item.questionCodeSystem;
               if (itemCodeSystem === 'LOINC') itemCodeSystem = serverFHIR.LOINC_URI;
-              fhirjs = LForms.fhirContext.getFHIRAPI(); // a fhir.js client
-
+              fhirClient = LForms.fhirContext;
               queryParams = {
-                type: 'Observation',
-                query: {
-                  code: itemCodeSystem + '|' + item.questionCode,
-                  _sort: '-date',
-                  _count: 5
-                }
+                code: itemCodeSystem + '|' + item.questionCode,
+                _sort: '-date',
+                _count: 5
               }; // only need one, but we need to filter out focus=true below
               // Temporarily disabling the addition of the focus search
               // parameter, because of support issues.  Instead, for now, we
@@ -5203,7 +5203,7 @@ function _typeof(obj) { "@babel/helpers - typeof"; if (typeof Symbol === "functi
               // returned.  Later, we might query the server to find out whether
               // :missing is supported.
               //if (LForms._serverFHIRReleaseID != 'STU3') // STU3 does not know about "focus"
-              //  queryParams.query.focus = {$missing: true}; // TBD -- sometimes :missing is not supported
+              //  queryParams.focus = {$missing: true}; // TBD -- sometimes :missing is not supported
 
               if (duration && duration.value && duration.code) {
                 // Convert value to milliseconds
@@ -5211,33 +5211,31 @@ function _typeof(obj) { "@babel/helpers - typeof"; if (typeof Symbol === "functi
 
                 if (result.status === 'succeeded') {
                   date = new Date(new Date() - result.toVal);
-                  queryParams.query._lastUpdated = 'gt' + date.toISOString();
+                  queryParams._lastUpdated = 'gt' + date.toISOString();
                 }
-              } // I'm not sure why, but fhirjs.search.then() returns an already
-              // resolved promise.  Wrap it in a Promise object.
+              } // TBD -delete              // I'm not sure why, but fhirjs.search.then() returns an already
+              // this too             // resolved promise.  Wrap it in a Promise object.
 
 
-              pendingPromises.push(new Promise(function (resolve, reject) {
-                fhirjs.search(queryParams).then(function (successData) {
-                  var bundle = successData.data;
+              pendingPromises.push(fhirClient.patient.request(_this2._buildURL(['Observation'], queryParams)).then(function (successData) {
+                var bundle = successData;
 
-                  if (bundle.entry) {
-                    var foundObs;
+                if (bundle.entry) {
+                  var foundObs;
 
-                    for (var j = 0, jLen = bundle.entry.length; j < jLen && !foundObs; ++j) {
-                      var obs = bundle.entry[j].resource;
+                  for (var j = 0, jLen = bundle.entry.length; j < jLen && !foundObs; ++j) {
+                    var obs = bundle.entry[j].resource;
 
-                      if (!obs.focus) {
-                        // in case we couldn't use focus:missing above
-                        serverFHIR.SDC.importObsValue(item, obs);
-                        if (item.value) foundObs = true;
-                        if (item.unit) lfData._setUnitDisplay(item.unit);
-                      }
+                    if (!obs.focus) {
+                      // in case we couldn't use focus:missing above
+                      serverFHIR.SDC.importObsValue(item, obs);
+                      if (item.value) foundObs = true;
+                      if (item.unit) lfData._setUnitDisplay(item.unit);
                     }
                   }
+                }
 
-                  resolve(item.questionCode); // code is not needed, but useful for debugging
-                });
+                return item.questionCode; // code is not needed, but useful for debugging
               }));
             }
           };
@@ -5245,7 +5243,7 @@ function _typeof(obj) { "@babel/helpers - typeof"; if (typeof Symbol === "functi
           for (var i = 0, len = this.itemList.length; i < len; ++i) {
             var duration;
             var itemCodeSystem;
-            var fhirjs;
+            var fhirClient;
             var queryParams;
             var result;
             var date;
@@ -7544,17 +7542,37 @@ function _typeof(obj) { "@babel/helpers - typeof"; if (typeof Symbol === "functi
      * @return a Promise that will resolve to the ValueSet expansion.
      */
     _fhirClientSearchByValueSetID: function _fhirClientSearchByValueSetID(valueSetID, fieldVal, count) {
-      var fhirClient = LForms.fhirContext.getFHIRAPI();
-      return fhirClient.search({
-        type: 'ValueSet/' + valueSetID + '/$expand',
-        query: {
-          _format: 'application/json',
-          count: count,
-          filter: fieldVal
-        }
-      }).then(function (response) {
-        return response.data;
-      });
+      var fhirClient = LForms.fhirContext;
+      return fhirClient.request(this._buildURL(['ValueSet', valueSetID, '$expand'], {
+        count: count,
+        filter: fieldVal
+      }));
+    },
+
+    /**
+     *  An internal function for safely constructing a URL based on a path
+     *  segments and query parameters.
+     * @param pathParts An array of path segments (not including the protocol).
+     *  These will be URL-encoded.
+     * @param queryParams (optional) A object of key/value pairs (string values)
+     *  for the query part of the URL.  Values be URL-encoded.
+     */
+    _buildURL: function _buildURL(pathParts, queryParams) {
+      // Would Use URL and URLSearchParams, except for the need to support IE11
+      // It looks like if we update babel, we could have a polyfill, but I don't
+      // want to revise the build configuration at the moment.
+      var url = pathParts.map(function (part) {
+        return encodeURIComponent(part);
+      }).join('/');
+
+      if (queryParams) {
+        // Assuming a single value per key, for now.
+        url += '?' + Object.keys(queryParams).map(function (k) {
+          return k + '=' + encodeURIComponent(queryParams[k]);
+        }).join('&');
+      }
+
+      return url;
     },
 
     /**
@@ -7568,7 +7586,7 @@ function _typeof(obj) { "@babel/helpers - typeof"; if (typeof Symbol === "functi
     _findValueSetIDAndSearch: function _findValueSetIDAndSearch(item, fieldVal, count) {
       // Fetch the ValueSet
       var failMsg = "Could not retrieve the ValueSet definition for " + item.answerValueSet;
-      var fhirClient = LForms.fhirContext.getFHIRAPI(); // Cache the lookup of ValueSet IDs, which should not change.  (A page
+      var fhirClient = LForms.fhirContext; // Cache the lookup of ValueSet IDs, which should not change.  (A page
       // reload will clear the cache.)
 
       if (!LForms._valueSetUrlToID) LForms._valueSetUrlToID = {};
@@ -7578,15 +7596,11 @@ function _typeof(obj) { "@babel/helpers - typeof"; if (typeof Symbol === "functi
         return this._fhirClientSearchByValueSetID(valueSetID, fieldVal, count);
       } else {
         var self = this;
-        return fhirClient.search({
-          type: 'ValueSet',
-          query: {
-            _format: 'application/json',
-            url: item.answerValueSet,
-            _total: 'accurate'
-          }
-        }).then(function (response) {
-          var data = response.data;
+        return fhirClient.request(this._buildURL(['ValueSet'], {
+          url: item.answerValueSet,
+          _total: 'accurate'
+        })).then(function (response) {
+          var data = response;
 
           if (data.total === 1) {
             var valueSetID = data.entry[0].resource.id;
@@ -7661,18 +7675,12 @@ function _typeof(obj) { "@babel/helpers - typeof"; if (typeof Symbol === "functi
             options.fhir = {
               search: function search(fieldVal, count) {
                 if (LForms.fhirCapabilities.urlExpandBroken) return self._findValueSetIDAndSearch(item, fieldVal, count);else {
-                  var fhirClient = LForms.fhirContext.getFHIRAPI();
-                  return fhirClient.search({
-                    type: 'ValueSet/$expand',
-                    query: {
-                      _format: 'application/json',
-                      count: count,
-                      filter: fieldVal,
-                      url: item.answerValueSet
-                    }
-                  }).then(function (successData) {
-                    return successData.data;
-                  }, function (errorData) {
+                  var fhirClient = LForms.fhirContext;
+                  return fhirClient.request(self._buildURL(['ValueSet/$expand'], {
+                    count: count,
+                    filter: fieldVal,
+                    url: item.answerValueSet
+                  })).catch(function (errorData) {
                     LForms.fhirCapabilities.urlExpandBroken = true; // HAPI does not support (maybe just because of a bug) $expand
                     // when both "url" and "filter" parameters are present.  In that
                     // case, it says, ""ValueSet contains include criteria with no